Have dedicated package for multi-tenancy support
