+++ server/src/main/java/org/spine3/server/command/CommandBus.java
@@ -31,8 +31,8 @@ import org.spine3.envelope.CommandEnvelope;
 import org.spine3.server.Statuses;
 import org.spine3.server.bus.Bus;
 import org.spine3.server.failure.FailureBus;
-import org.spine3.server.storage.TenantAware;
-import org.spine3.server.storage.TenantIndex;
+import org.spine3.server.tenant.TenantAware;
+import org.spine3.server.tenant.TenantIndex;
 import org.spine3.type.CommandClass;
 import org.spine3.users.TenantId;
 import org.spine3.util.Environment;
