+++ server/src/main/java/org/spine3/server/tenant/DefaultTenantRepository.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2017, TeamDev Ltd. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package org.spine3.server.tenant;
+
+import com.google.protobuf.Timestamp;
+import org.spine3.protobuf.Timestamps2;
+import org.spine3.users.TenantId;
+
+/**
+ * Default implementation of {code TenantRepository} that stores timestamps
+ * of tenant ID registration.
+ *
+ * @author Alexander Yevsyukov
+ */
+public final class DefaultTenantRepository extends TenantRepository<Timestamp,
+                                                              DefaultTenantRepository.Entity> {
+
+    public static class Entity extends TenantRepository.Entity<Timestamp> {
+        protected Entity(TenantId id) {
+            super(id);
+        }
+
+        @Override
+        public Timestamp getDefaultState() {
+            return Timestamps2.getCurrentTime();
+        }
+    }
+}
