+++ server/src/main/java/org/spine3/server/tenant/TenantAware.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.server.storage;
+package org.spine3.server.tenant;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Optional;
