+++ server/src/test/java/org/spine3/server/command/AbstractCommandBusTestSuite.java
@@ -32,8 +32,8 @@ import org.spine3.client.CommandFactory;
 import org.spine3.envelope.CommandEnvelope;
 import org.spine3.server.event.EventBus;
 import org.spine3.server.failure.FailureBus;
-import org.spine3.server.storage.DefaultTenantRepository;
 import org.spine3.server.storage.memory.InMemoryStorageFactory;
+import org.spine3.server.tenant.DefaultTenantRepository;
 import org.spine3.test.TestCommandFactory;
 import org.spine3.test.command.CreateProject;
 import org.spine3.test.command.event.ProjectCreated;
